fix router folder and add loader
