+++ src/Pages/LoginPage/Login.css
@@ -1,0 +1,19 @@
+.waitingMessage-enter {
+  opacity: 0;
+  transform: translateX(-100%);
+}
+.waitingMessage-enter-active {
+  opacity: 1;
+  transform: translateX(0);
+  transition: all 250ms linear;
+}
+
+.waitingMessage-exit {
+  opacity: 1;
+  transform: translateX(0);
+}
+.waitingMessage-exit-active {
+  opacity: 0;
+  transform: translateX(-100%);
+  transition: all 250ms linear;
+}
