+++ src/Pages/SingUpPage/SingUpPage.module.scss
@@ -1,0 +1,5 @@
+.waitingMessage {
+  width: 80px;
+  margin-left: auto;
+  margin-right: auto;
+}
